refactor: move AlertVariant type to a dedicated file and update imports
